* develop: (926 commits)
  add function _fetch_template()
  forum_viewthread.php
  [desc] function strreplace_strip_split
  fix - part.3
  fix - part.2
  fix
  變更主題列表標題的預設顯示模式
  bugfix
  fix
  update
  split
  [desc] function get_seosetting
  [desc] function get_title_page
  [desc] 主題標題 - 第2頁
  在論壇 › 論壇管理的導航顯示目前的版塊名稱
  add $navigation
  fix
  [desc] 用來解決升級轉換後無法保存舊有的 onlinerecord 紀錄問題
  BUG:升級轉換後無法保存 onlinerecord 會被強制性覆寫
  use updatecache('onlinerecord');
  ...

+++ discuzx/utility/convert/source/d7.2_x2.0/optionvalue.php
@@ -22,25 +22,32 @@ $_sorts = getgpc('_sorts');
 if(empty($_sorts)) {
 	$_sorts = array();
 	// 搜尋相關的 TABLE
-	$query = $db_source->query("SHOW TABLES LIKE '{$config[source][tablepre]}optionvalue%'");
-	while($row = $db_source->fetch_row($query)) {
+	$query = mysql_list_tables($db_source->config['dbname'], $db_source->curlink);
+	$num_rows = mysql_num_rows($query);
+	for ($i = 0; $i < $num_rows; $i++) {
+		$row = mysql_tablename($query, $i);
+
+		if (strpos($row, $db_source->config['tablepre'].'optionvalue') !== 0) continue;
+
 		$tabledump = '';
 		$sortid = 0;
 		// 檢查是否為正確想要搜尋的 TABLE
-		$sortid = str_replace($config[source][tablepre].'optionvalue', '', $row[0]);
-		if ($sortid && $sortid == intval($sortid)) {
+		$sortid = str_replace($db_source->config['tablepre'].'optionvalue', '', $row);
+		if (!empty($sortid) && $sortid == intval($sortid) && is_numeric($sortid)) {
+			$sortid == intval($sortid);
+
 			// 複製來源 TABLE 結構
-			if ($tabledump = _sqldumptablestruct($config[source][tablepre].'optionvalue'.$sortid, $db_source)) {
+			if ($tabledump = _sqldumptablestruct($db_source->config['tablepre'].'optionvalue'.$sortid, $db_source)) {
 				// 刪除已存在的目標 TABLE
 				$db_target->query("DROP TABLE IF EXISTS ".$table_target.$sortid);
 				// 轉換 TABLE 表的建立語法 將來源 TABLE 名稱取代為目標 TABLE 名稱
-				$tabledump = str_replace('CREATE TABLE `'.$config[source][tablepre].'optionvalue'.$sortid.'`', 'CREATE TABLE `'.$config[target][tablepre].'forum_optionvalue'.$sortid.'`', $tabledump);
+				$tabledump = str_replace('CREATE TABLE `'.$db_source->config['tablepre'].'optionvalue'.$sortid.'`', 'CREATE TABLE `'.$db_target->config['tablepre'].'forum_optionvalue'.$sortid.'`', $tabledump);
 				$db_target->query($tabledump);
 			}
 			$_sorts[] = $sortid;
 		}
 	}
-	
+
 	unset($sortid, $tabledump, $query );
 } else {
 	// 將接收的 $_sorts 轉換為 Array
@@ -49,16 +56,19 @@ if(empty($_sorts)) {
 
 $nextid = -1;
 // 取出陣列中第一個值作為目前要處理的 sortid
-$_sortid = $_sorts[0];
+if ($_sortid = $_sorts[0]) {
 
-$query = $db_source->query("SELECT * FROM $table_source$_sortid WHERE tid>'$start' ORDER BY tid LIMIT $limit");
-while ($row = $db_source->fetch_array($query)) {
-	$nextid = $row['tid'];
+	$query = $db_source->query("SELECT * FROM $table_source$_sortid WHERE tid>'$start' ORDER BY tid LIMIT $limit");
+	while ($row = $db_source->fetch_array($query)) {
+		$nextid = $row['tid'];
 
-	$row  = daddslashes($row, 1);
-	$data = implode_field_value($row, ',', db_table_fields($db_target, $table_target.$_sortid));
-	$db_target->query("INSERT INTO $table_target$_sortid SET $data");
+		$row  = daddslashes($row, 1);
+		$data = implode_field_value($row, ',', db_table_fields($db_target, $table_target.$_sortid));
+		$db_target->query("INSERT INTO $table_target$_sortid SET $data");
+	}
+
 }
+
 // 如果沒有查詢任何資料
 if($nextid < 0) {
 	// 取出陣列的第一個值
